+++ docs/project/changelog.md
@@ -98,6 +98,10 @@ substitutions:
 - {{ Fix }} Shared libraries with version suffix are now handled correctly.
   {pr}`3154`
 
+- Added a new CLI command `pyodide sekeleton` which creates a package build recipe.
+  `pyodide-build mkpkg` will be replaced by `pyodide sekeleton pypi`.
+  {pr}`3175`
+
 - {{ Fix }} Fixed bug in `split` argument of {any}`repr_shorten`. Added {any}`shorten` function.
   {pr}`3178`